feat: US-03 - Split sparse_index/search.rs into modular search/ directory
Split the 695-line search.rs into 3 focused modules:
- search/scoring.rs: term posting structures, scoring primitives, result extraction
- search/strategy.rs: MaxScore DAAT and linear scan search strategies
- search/mod.rs: orchestrator with public API, constants, and tests
